arm64: dts: freescale: imx95-toradex-smarc: reserve memory for security enclave
The security enclave firmware runs on the Cortex-M33 core, and the
corresponding Linux driver on the Cortex-A core expects responses to
requests to be returned via a shared memory region in DRAM.

Since the Cortex-M33 cannot access the entire DDR memory space, reserve
a Cortex-M33 accessible memory region for the security enclave firmware.

The address match the same address used by U-Boot.

Upstream-Status: Pending
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>