arm64: dts: freescale: imx95-aquila: Add SPI1, SPI2 and on-module TPM
Add Aquila SPI1, SPI2 and on-module TPM. The TPM is connected as the
second device on SPI2 to avoid conflicts with the spidev overlay. Also,
its chip select and interrupt line (falling edge) are provided via the
I2C I/O expander at address 0x21. A label was added to the TPM node to
avoid conflicts with carrier board TPM, or in case its properties
need to be overridden.

Upstream-Status: Pending
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>