arm64: dts: freescale: imx95-aquila-dev: Add board I2C devices
Add I2C devices present on the development board:
- Fan controller AMC6821 (0x18)
- Audio codec WM8904 (0x1a)
- Current measurement INA226 (0x41)
- Temperature sensor TMP1075 (0x4f)
- EEPROM ST24C02 (0x57)

An alias and label were added to the EEPROM to make it easier to
differentiate it from the SoM EEPROM (0x50).

For the fan controller, only a default configuration was added for now
(setting the fan speed to maximum). The correct thermal configurations
must be provided when thermal zones are added.

For the audio codec, the remaining audio configuration (sound card)
still must be implemented.

Upstream-Status: Pending
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>