drm/bridge: imx: fix mixed module-builtin object
commit 7eb790eb048b ("Revert "drm/bridge: imx: fix mixed module-builtin
object"") reverted that the imx-ldb-helper can be used as module and
with later commits stopped using the file in downstream for i.MX
8QXP/8QM).

The commit 0b608ae89fac ("LF-10583-2: drm: bridge: imx: Add i.MX95 LVDS
Display Bridge(LDB) driver") used the file again for i.MX 95 without
readding the posibility to use it as a module.

Revert the Revert so that imx-ldb-helper can be used from a module
independent of it being built as a module or built in.

Submitted to NXP downstream.

Upstream-Status: Submitted [https://github.com/nxp-imx/linux-imx/pull/34/commits]
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>