NVIDIA: SAUCE: net: phy: realtek: Fix RTL8211F suspend
BugLink: https://bugs.launchpad.net/bugs/2127135

When integrating WoL support for the RTL8211F device, a duplicated
suspend entry was added for this device. The difference between the
rtl821x_suspend() and the genphy_suspend() functions is that the
rtl821x_suspend() will only call genphy_suspend() and disable clocks
if WoL is not enabled. Therefore, remove the duplicated suspend entry
that calls genphy_suspend() to ensure the PHY is not suspend if WoL is
enabled.

http://nvbugs/200562291

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-by: Narayana Reddy P <narayanr@nvidia.com>
Reviewed-by: Brad Griffis <bgriffis@nvidia.com>
Signed-off-by: Prathamesh Shete <pshete@nvidia.com>
Acked-by: Abdur Rahman <abdur.rahman@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>