dt-bindings: arm: fsl: Add Lino iMX93
Lino iMX93 is a SoM using the NXP i.MX93 SoC, it connects with a carrier
board with two board to board connectors.

The Lino SoM can also be plugged onto an adapter board that converts it to
the Toradex Verdin family pinout, and with that it can be
plugged into any Verdin compatible carrier board.

Upstream-Status: Pending
Link: https://www.toradex.com/computer-on-modules/lino-arm-family/nxp-imx93
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>