phy: freescale: kconfig: add typec dependency for imx8mq usb phy
The iMX8MQ USB PHY driver uses various functions of the typec driver,
which are only available if the typec subsystem is selected.
Without a proper Kconfig dependecy, the driver can be built in
configurations where the required typec functions are missing,
leading to build failures.

This is not relevant for mainline as it was introduced with commit
b58f0f86fd61 ("phy: fsl-imx8mq-usb: add tca function driver for imx95")

Upstream-Status: Inappropriate [Other]
Signed-off-by: Franz Schnyder <franz.schnyder@toradex.com>