arm64: dts: imx95: fix mali thermal zone
The mali binding includes the thermal-zone string property which
expects the node name of the thermal zone assosiated with the GPU.

We changed the thermal zone node name without changing the thermal-zone
property in the mali node. This results in:

| [    5.354741] Error -19 getting thermal zone 'ana', not yet ready?
| [    5.354765] mali 4d900000.gpu: recalculation of power model mali-simple-power-model returned error -517
| [    5.364256] mali 4d900000.gpu: IPA initialization failed
| [    5.369592] mali 4d900000.gpu: Continuing without devfreq

Follow the node name change and the thermal code of the mali driver
prints the following:

| [    8.620732] mali 4d900000.gpu: Using configured power model mali-tvax-power-model, and fallback mali-simple-power-model
| [    8.636987] mali 4d900000.gpu: OPP 0 : opp_freq=500000000 core_mask=1
| [    8.636995] mali 4d900000.gpu: OPP 1 : opp_freq=800000000 core_mask=1
| [    8.651876] mali 4d900000.gpu: OPP 2 : opp_freq=1000000000 core_mask=1

Mainline has no Mali GPU support for i.MX 95.

Upstream-Status: Inappropriate [Downstream only]
Fixes: 6bea215e4156 ("arm64: dts: imx95: align thermal-zone cpu node names with upstream")
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>